+++ utils.py
@@ -1,13 +1,24 @@
 from functools import wraps
 from flask import jsonify, request
-import documents.state_machine as state_machine
+import inspect
+from config import config
+import jwt
+
+def token_valid_check(auth_header):
+    auth_token = auth_header.split(' ')[1]
+    if auth_token == "" or auth_token == None:
+        raise Exception({'message':'Authorization token empty'})
+    decoded_token = jwt.decode(auth_token, config['ACCESS_TOKEN_SECRET'], algorithms=["HS256"])
+    return decoded_token
 
 def authenticate_user(func):
     @wraps(func)
-    def wrapper():
+    def wrapper(*args, **kwargs):
         auth_header = request.headers.get('Authorization')
-        payload = state_machine.token_valid_check(auth_header)
-        return func(payload)
+        payload = token_valid_check(auth_header)
+        if 'payload' in inspect.getfullargspec(func).args:
+            kwargs['payload'] = payload
+        return func(*args, **kwargs)
     return wrapper
 
 def server_error_check(func):
